pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ragpoisoner"
version = "0.1.0"
description = "RAG Corpus Poisoning Toolkit — AI red teaming for RAG systems"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
keywords = ["security", "red-team", "RAG", "LLM", "AI security", "pentesting"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Information Technology",
    "Topic :: Security",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "click>=8.1",
    "chromadb>=0.5",
    "sentence-transformers>=3.0",
    "requests>=2.31",
    "jinja2>=3.1",
    "numpy>=1.26",
    "scipy>=1.11",
    "rich>=13.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = ["pytest>=7", "pytest-cov", "black", "ruff"]
yaml = ["pyyaml>=6.0"]

[project.scripts]
ragpoisoner = "ragpoisoner.cli:main"

[project.urls]
Homepage = "https://github.com/x0verFl0w/ragpoisoner"
"Bug Tracker" = "https://github.com/x0verFl0w/ragpoisoner/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["ragpoisoner*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]
